% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{AsmannTissues}
\alias{AsmannTissues}
\title{Human Tissue Experiment (Asmann et al.)}
\format{List with 3 elements:
\describe{
  \item{Expression}{Integer matrix: Features by samples matrix with integers representing raw counts}
  \item{Design}{Dataframe: Study design}
  \item{Annotation}{Dataframe: Additional info about features}
  The three objects are already in the same order, but can be matched using row/column names.
}}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/22496456}
}
\usage{
AsmannTissues
}
\description{
RNA-Seq data originally from RECount. For more information, see reference from URL.
}
\keyword{datasets}

