% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{copy_airway_bamfiles}
\alias{copy_airway_bamfiles}
\title{Airway dexamethasone experiment BAM files}
\source{
\url{http://bioconductor.org/packages/release/data/experiment/html/airway.html}
}
\usage{
copy_airway_bamfiles()
}
\description{
Creates a directory "airway/" in the workind directory and copies example BAM-files into this folder.
}
\details{
RNA-Seq data originally from bioconductor. For more information, see reference from URL.
}
\author{
Malte Thodberg
}

